Complete rewrite of the script
What I wrote this script for was to have accurate percentages while
changing the brightness, but unfortunately, the previous implementation
didn't have it. I've improved it, and I've changed pretty much the whole
code, and made it more efficient and faster to run. And the code is way
shorter than before.

Previously the code was slightly slow (noticeably), now its more
efficient